# backend/api/openapi/melete.yaml
openapi: 3.0.3
info:
  title: Melete API
  version: 0.1.0
  description: |
    题库驱动的学习平台。

    设计要点：**答案不是单一字段，而是一组带来源的「主张」**（`AnswerClaim`）。
    884 道有对照数据的题里 38% 存在题库标注与社区投票不一致，
    界面上并列呈现各方主张本身就是最好的学习材料。
servers:
  - url: /api/v1

# 认证：Bearer access token（见 docs/design/active/deployment.md §0）。
#
# 本 API 对公网暴露（web 与 iOS 都是平等客户端），因此不存在「可信的调用方」——
# 曾有过的服务间共享密钥已移除：原生 App 反编译即得密钥，它挡不住任何人。
# 身份只由 access token 的签名保证，accountId 取自其 sub。
security:
  - accessToken: []
...
paths:
  /auth/password:
    post:
      tags: [auth]
      summary: 用户名密码登录
      operationId: passwordLogin
      security: []          # 登录前尚无 token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: 认证通过
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          description: 用户名或密码错误
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/sso:
    post:
      tags: [auth]
      summary: 用 Akasha id_token 换取本 API 的 token
      description: |
        web 与 iOS 各自完成 OIDC 流程后，把拿到的 id_token 交给本端点。

        **id_token 由本 API 自己验签**（Akasha JWKS + issuer + audience）——
        绝不接受客户端自称的 sub：那等于任何人都能登录任意账号。
      operationId: ssoExchange
      security: []          # 登录前尚无 token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SsoExchange" }
      responses:
        "200":
          description: 账号已确立
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          description: id_token 无效
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: 用 refresh token 换新的 access token
      operationId: refreshToken
      security: []          # access 已过期，此处只认 refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: 已换发
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          description: refresh token 无效、已过期或已吊销
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /auth/sessions/revoke-others:
    post:
      tags: [auth]
      summary: 登出其它设备（保留当前这台）
      description: |
        ⭐ 「保留哪一条」取自当前 access token 的 sid，⛔ 不接受客户端指定 ——
        否则任何人都能构造一个「保留别人的会话、踢掉我的」的请求。
        ⚠️ 若当前 token 没有 sid（阶段 3 之前签发的旧票），返回 409 而
        ⛔ 不退化成「登出全部」—— 那会把一次误操作放大成全员掉线。
      operationId: revokeOtherSessions
      responses:
        "200":
          description: 已登出其它设备
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked: { type: integer, description: 被登出的会话数 }
        "409":
          description: 当前令牌不含会话标识，请重新登录后再试
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /attempts:
    post:
      tags: [study]
      summary: 记录一次作答（对错由服务端按参考答案判定）
      operationId: recordAttempt
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttemptInput" }
      responses:
        "200":
          description: 已记录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttemptResult" }
        "404": { $ref: "#/components/responses/NotFound" }
  /me/progress:
    get:
      tags: [study]
      summary: 我的学习进度总览
      operationId: getMyProgress
      parameters:
        - $ref: "#/components/parameters/BankQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Progress" }
  /me/tag-stats:
    get:
      tags: [study]
      summary: 按标签的正确率（「我哪里不会」）
      description: |
        `attempt ⋈ question_tag ⋈ tag` 聚合。每题只取**最近一次**作答参与统计 ——
        否则反复重做同一道错题会把正确率越刷越低，与直觉相反。
      operationId: getMyTagStats
      parameters:
        - $ref: "#/components/parameters/BankQuery"
        - name: type
          in: query
          required: true
          schema: { type: string, enum: [domain, topic, concept] }
        - name: minAttempts
          in: query
          description: 至少做过几道才纳入（样本太小的标签正确率没有意义）
          schema: { type: integer, minimum: 1, default: 3 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TagStat" }
...
  /me/recent:
    get:
      tags: [study]
      summary: 最近的学习会话（跨题库）
      description: |
        从 attempt 推导：同一题库、同一 context、相邻间隔 ≤ 30 分钟的连续作答算一次会话。
        不存会话表 —— 会话是对事实的解释，不是事实。
      operationId: getMyRecent
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 5 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StudySession" }
  /me/resume:
    get:
      tags: [study]
      summary: 继续学习（上次刷到哪）
      description: 从 attempt 推导，不额外维护断点表 —— 少一处要保持一致的状态。
      operationId: getMyResume
      parameters:
        - $ref: "#/components/parameters/BankQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resume" }
...
  /banks/{slug}:
    get:
      tags: [bank]
      summary: 题库详情与统计
      operationId: getBank
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BankDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /banks/{slug}/tags:
    get:
      tags: [bank]
      summary: 题库的标签列表（用于筛选与正确率热图）
      operationId: listBankTags
      parameters:
        - $ref: "#/components/parameters/Slug"
        - name: type
          in: query
          description: 只返回某一类标签
          schema: { type: string, enum: [domain, topic, concept] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
        "404": { $ref: "#/components/responses/NotFound" }
  /banks/{slug}/questions:
    get:
      tags: [question]
      summary: 题目列表（分页 + 标签筛选）
      operationId: listQuestions
      parameters:
        - $ref: "#/components/parameters/Slug"
        - name: tag
          in: query
          description: 标签 id，可重复；多个标签取交集
          schema:
            type: array
            items: { type: integer, format: int64 }
        - name: contested
          in: query
          description: 只返回存在答案分歧的题（题库标注 ≠ 社区投票）
          schema: { type: boolean }
        - name: enriched
          in: query
          description: 只返回已完成 AI 富化的题
          schema: { type: boolean }
        - name: mode
          in: query
          description: |
            学习模式过滤（全部需要会话，账号取自 JWT 而非请求参数）：
            wrong = 最近一次作答是错的；unsure = 最近一次自评 rating ≤ 2；
            unseen = 从未作答；
            due = FSRS 复习队列，卡片已到期。

            ⚠️ due 与 unseen 互斥且不重叠：没做过的题没有卡片，不算「到期」。
            两者混进一个数字会让「今天要复习 300 题」失去意义。
            ⚠️ due 按【到期时间】升序返回，其余模式按原题号 —— 复习队列不是浏览列表。
          schema: { type: string, enum: [wrong, unsure, unseen, due] }

        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /questions/{id}:
    get:
      tags: [question]
      summary: 单题详情（题干 + 选项 + 各方答案主张 + 解析 + 标签）
      operationId: getQuestion
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        本 API 签发的 access token（HS256，TTL 1 小时，对齐 geass-v3 规范）。
        过期后用 refresh token 调 /auth/refresh 换新的。
  parameters:
    BankQuery:
      name: bank
      in: query
      description: 题库 slug；缺省 = 当前题库（现阶段为第一个题库）
      schema: { type: string }
    Slug:
      name: slug
      in: path
      required: true
      description: 题库 slug，如 aws-saa-c03
      schema: { type: string }
  responses:
    TooManyRequests:
      description: 登录尝试过于频繁
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Progress:
      type: object
      required: [bankSlug, questionCount, seenCount, correctCount, wrongCount, unsureCount, attemptCount, dueCount]
      properties:
        bankSlug:     { type: string }
        questionCount: { type: integer, description: 题库总题数 }
        seenCount:    { type: integer, description: 做过的题数（去重） }
        correctCount: { type: integer, description: 最近一次作答为对的题数 }
        wrongCount:   { type: integer, description: 最近一次作答为错的题数 }
        unsureCount:  { type: integer, description: 最近一次自评 ≤2 的题数 }
        attemptCount: { type: integer, description: 总作答次数（含重做） }
        dueCount:
          type: integer
          description: |
            FSRS 复习队列里已到期的题数。
            ⚠️ **不含从没做过的题** —— 那些题没有卡片，属于「没做过」入口。
            两者混进一个数字，「今天要复习 300 题」就失去意义了，
            而那正是间隔重复最劝退的失败模式。
        lastActiveAt: { type: string, format: date-time }
    TagStat:
      type: object
      required: [tagId, type, value, total, correct, rate]
      properties:
        tagId:   { type: integer, format: int64 }
        type:    { type: string, enum: [domain, topic, concept] }
        value:   { type: string }
        i18n:
          type: object
          additionalProperties: { type: string }
        total:   { type: integer, description: 该标签下做过的题数 }
        correct: { type: integer }
        rate:    { type: number, format: float, description: 正确率 0-100 }
    Resume:
      type: object
      description: |
        「继续学习」的两条轨道，全部从 attempt 推导：
        · sequential —— 顺序进度：题号最小的没做过的题，无需存状态
        · focus —— 上次专项：最近一条 context.mode ∉ {unseen, all} 的作答的出处；从未做过专项时缺省
      required: [bankSlug, sequential]
      properties:
        bankSlug:   { type: string }
        sequential: { $ref: "#/components/schemas/SequentialCursor" }
        focus:      { $ref: "#/components/schemas/FocusCursor" }
...
    StudySession:
      type: object
      required: [bankSlug, bankName, context, label, startedAt, endedAt, count, correct]
      properties:
        bankSlug:  { type: string }
        bankName:  { type: string }
        context:   { $ref: "#/components/schemas/DrillContext" }
        label:     { type: string, description: mode=tag 时为标签名，否则为 mode }
        tag:       { $ref: "#/components/schemas/Tag" }
        startedAt: { type: string, format: date-time }
        endedAt:   { type: string, format: date-time }
        count:     { type: integer }
        correct:   { type: integer }
        firstNo:   { type: integer, description: 顺序刷时的起止题号，便于显示 "#400–#413" }
        lastNo:    { type: integer }
    SequentialCursor:
      type: object
      required: [doneCount, totalCount]
      description: questionId 缺省 = 题库已全部做过一遍
      properties:
        questionId: { type: integer, format: int64 }
        externalNo: { type: integer }
        stem:       { type: string }
        doneCount:  { type: integer, description: 做过的题数（去重） }
        totalCount: { type: integer, description: 题库总题数 }
        lastAt:     { type: string, format: date-time, description: 上次顺序作答时间 }
    FocusCursor:
      type: object
      required: [context, label, total, lastAt]
      properties:
        context: { $ref: "#/components/schemas/DrillContext" }
        label:   { type: string, description: mode=tag 时为标签 value，否则为 mode 名 }
        tag:     { $ref: "#/components/schemas/Tag" }
        done:    { type: integer, description: 集合内做过的题数；只有 tag / contested 有意义 }
        total:   { type: integer, description: 集合当前大小 }
        lastAt:  { type: string, format: date-time }
    DrillContext:
      type: object
      description: 一次作答的出处 —— 用户是从哪个入口做的这道题
      required: [mode]
      properties:
        mode:  { type: string, enum: [all, unseen, wrong, unsure, contested, tag, due] }
        tagId: { type: integer, format: int64, description: mode=tag 时必填 }
    AttemptInput:
      type: object
      required: [questionId, chosen, rating]
      properties:
        questionId: { type: integer, format: int64 }
        chosen:     { type: string, description: 排序后的字母集合，如 "AB" }
        rating:
          type: integer
          minimum: 1
          maximum: 4
          description: FSRS 标准四键自评 1=Again 2=Hard 3=Good 4=Easy
        durationMs: { type: integer }
        context:    { $ref: "#/components/schemas/DrillContext" }
    AttemptResult:
      type: object
      required: [correct]
      properties:
        correct: { type: boolean, description: 服务端按参考答案判定 }
        reference: { $ref: "#/components/schemas/Reference" }
        schedule: { $ref: "#/components/schemas/ScheduleResult" }
    ScheduleResult:
      type: object
      description: |
        这次作答之后的记忆调度结果（FSRS）。

        **自评的可信度纠正**：`rating` 是用户按下的键、原样落进 attempt 表；
        `effectiveRating` 是实际喂给调度器的值。两者不同时 `reasons` 说明为什么 ——
        ⛔ 界面必须把理由显示出来，不得偷偷改调度。
        这与三方答案主张并列展示是同一条哲学：不替学习者下结论，把分歧摆出来。

        当前的纠正规则：
        · 答错却自评「掌握 / 轻松」→ 压到 1（作答顺序是【揭晓之后】才自评，
          答案就在眼前，不存在「只是没算准」的解释空间）
        · 答对但用时短到读不完题面 → 压到 2（答对了，但这次答对提供不了「你会」的证据）
        · 没有任何答案主张的题（题库里 6 道）→ 不纠正，判不了对错就不拿对错说事
      required: [rating, effectiveRating, reasons, due, state]
      properties:
        rating:          { type: integer, minimum: 1, maximum: 4, description: 用户按下的自评键 }
        effectiveRating: { type: integer, minimum: 1, maximum: 4, description: 实际用于调度的评分，只会 ≤ rating }
        reasons:
          type: array
          description: 纠正理由，未纠正时为空数组
          items: { type: string }
        due:   { type: string, format: date-time, description: 这道题下次该出现的时间（UTC） }
        state: { type: string, enum: [new, learning, review, relearning], description: 调度后的卡片状态 }
    Reference:
      type: object
      description: |
        判对错用的参考答案。优先级 ai_verdict > community_vote > bank_label ——
        AI 是唯一看过全部信息并给出理由的来源；题库标注是四个来源里
        最不该被无条件信任的那个（38% 与社区投票不一致）。
        仅用于给出反馈信号，界面仍并列展示全部主张。
      required: [answer, source]
      properties:
        answer: { type: string }
        source: { type: string, enum: [bank_label, community_vote, ai_verdict, user_note] }
    Credentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        deviceInfo: { type: string, description: 设备标识，用于「我的登录设备」 }
    SsoExchange:
      type: object
      required: [idToken]
      properties:
        idToken: { type: string, description: Akasha 签发的 id_token（由本 API 验签） }
        deviceInfo: { type: string, description: 设备标识，用于「我的登录设备」 }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
...
    TokenPair:
      type: object
      description: |
        access 是 JWT（不落库，TTL 1h）；refresh 是不透明随机串（落库，可吊销）。
        对齐 geass-v3：refresh 不做成 JWT —— 生命周期以月计的凭证必须能撤回。
      required: [accessToken, refreshToken, expiresIn, accountId, display]
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        expiresIn:    { type: integer, description: access token 剩余秒数（非绝对时间，免受客户端时钟偏差影响） }
        # ⚠️ 2026-09-07 breaking change：integer/int64 → string（canonical UUID）。
        # 账号 id 从 BIGINT 自增换成 UUIDv7（cross-exam 005 §22.2）——
        # 理由是 TiDB 跳号严重，且 v7 时间有序避免 v4 的索引页分裂。
        # ⛔ 消费方（web / iOS）必须同步改，⚠️ 旧 token 里的数字 sub 会被拒。
        # ⛔ 刻意【不写 format: uuid】：那会让 oapi-codegen 产出
        # openapi_types.UUID，于是 httpapi 层要多一次 UUID↔string 转换 ——
        # 而「编解码只有一处」是本项目的硬规则（见 internal/localauthx/uuid.go）。
        # 契约上它就是一个字符串；形态由 pattern 说明，⛔ 不由类型强加。
        accountId:
          type: string
          pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
          description: 账号 id（canonical UUIDv7 文本）
        display:      { type: string }
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Bank:
      type: object
      required: [id, slug, name, locale, kind, meta]
      properties:
        id:          { type: integer, format: int64 }
        slug:        { type: string, example: aws-saa-c03 }
        name:        { type: string }
        description: { type: string }
        locale:      { type: string, example: zh }
        kind:        { type: string, enum: [cert, custom] }
        meta:        { $ref: "#/components/schemas/BankMeta" }
...
    BankDetail:
      allOf:
        - $ref: "#/components/schemas/Bank"
        - type: object
          required: [stats]
          properties:
            stats: { $ref: "#/components/schemas/BankStats" }
    BankStats:
      type: object
      description: 题库的内容侧统计（与用户无关）
      required: [questionCount, enrichedCount, contestedCount]
      properties:
        questionCount:  { type: integer, description: 题目总数 }
        enrichedCount:  { type: integer, description: 已完成 AI 富化的题数 }
        contestedCount: { type: integer, description: 题库标注与社区投票不一致的题数 }
    Tag:
      type: object
      description: 通用 (type, value) 结构，不硬编码任何特定题库的分类体系
      required: [id, type, value]
      properties:
        id:    { type: integer, format: int64 }
        type:  { type: string, enum: [domain, topic, concept] }
        value: { type: string, example: "S3" }
        i18n:
          type: object
          additionalProperties: { type: string }
        questionCount: { type: integer, description: 挂在该标签下的题数 }
    Choice:
      type: object
      required: [label, body]
      properties:
        label: { type: string, example: "A" }
        body:  { type: string }
...
    QuestionSummary:
      type: object
      required: [id, externalNo, stem, kind, pickCount, contested, enriched]
      properties:
        id:         { type: integer, format: int64 }
        externalNo: { type: integer, description: 原题号，用于溯源 }
        stem:       { type: string }
        kind:       { type: string, enum: [single, multi] }
        pickCount:  { type: integer }
        contested:  { type: boolean, description: 题库标注与社区投票是否不一致 }
        enriched:   { type: boolean, description: 是否已完成 AI 富化 }
    QuestionDetail:
      allOf:
        - $ref: "#/components/schemas/QuestionSummary"
        - type: object
          required: [bankSlug, choices, claims, explanations, tags]
          properties:
            bankSlug:  { type: string }
            reference: { $ref: "#/components/schemas/Reference" }
            dataIssue:
              type: string
              description: 题目本身有缺陷时的说明（题库脏点），正常不出现
            choices:
              type: array
              items: { $ref: "#/components/schemas/Choice" }
            claims:
              type: array
              items: { $ref: "#/components/schemas/AnswerClaim" }
            explanations:
              type: array
              items: { $ref: "#/components/schemas/Explanation" }
            tags:
              type: array
              items: { $ref: "#/components/schemas/Tag" }
            warnings:
              type: array
              description: P0 解析阶段的告警，不静默丢弃
              items: { type: string }
    QuestionPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QuestionSummary" }
        total:  { type: integer }
        limit:  { type: integer }
        offset: { type: integer }